cloud_function/requirements.txt
# Google Cloud dependencies
google-cloud-storage==2.14.0
google-cloud-bigquery==3.15.0
functions-framework==3.5.0

# Bioinformatics libraries
biopython==1.81
pysam==0.21.0

# Utilities
numpy==1.24.3

